Did some more refactoring handling moving over to indexed textures instead, think I am about done with that now and can move on. I also went through and fixed some of the different warnings visual studio were complaining about, like saying a float = GLuint, so I casted them over, which isn't really that nice, but I am aware of those warnings, so I am mainly casting to shut the compiler up, which isn't really nice

+++ Graphics Exam 2015/InputHandler.cpp
@@ -214,8 +214,11 @@ void InputHandler::handleMouse(SDL_Event& eventHandler, std::queue<GameEvent>& e
             std::cout << "Button wheel" << std::endl;
             break;
         case SDL_MOUSEMOTION:
-            mousePosition.x = eventHandler.motion.x;
-            mousePosition.y = eventHandler.motion.y;
+                              // Not really nice to cast here
+                              // But Visual studio generates warnings for possible loss of data, which makes it hard to see other warnings
+                              // so as I am aware of what I am doing, I cast it instead.
+            mousePosition.x = static_cast<float>(eventHandler.motion.x);
+            mousePosition.y = static_cast<float>(eventHandler.motion.y);
             action = ActionEnum::MOUSEMOTION;
             break;
     }
